Renames and moves the LogLevel class.
